feat: implement syncExternal functionality and update effect handling in store
